changed price field from integer to decimal

+++ LittleLemonFS/models.py
@@ -10,7 +10,7 @@ class Booking(models.Model):
     
 class Menu(models.Model):
    name = models.CharField(max_length=200) 
-   price = models.IntegerField(null=False) 
+   price = models.DecimalField(null=False, decimal_places=2, max_digits=6) 
    menu_item_description = models.TextField(max_length=1000, default='') 
 
    def __str__(self):
